Fixed yet another compiler error
Former-commit-id: 78b1b238a0e9ec94da750c046111e68b44c7a45d

+++ src/librapid/modified/float16/float16.hpp
@@ -1074,10 +1074,9 @@ namespace librapid {
 	namespace internal {
 		LR_INLINE extended::float16_t copySign(const extended::float16_t &mag,
 											   const extended::float16_t &sign) noexcept {
-			return extended::copysign(mag, sign);
+			return extended::copysign(static_cast<float>(mag), static_cast<float>(sign));
 		}
 	}
-
 } // namespace librapid
 
 // Provide {fmt} printing capabilities
